requirements.txt
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
openai==1.65.2
PyAudio==0.2.14
pyttsx3==2.98
requests==2.32.3
SpeechRecognition==3.14.1